nosql/ana_report/requirements.txt
fastapi
uvicorn
psycopg2-binary
pandas
python-dotenv
numpy
plotly
matplotlib
fpdf
jinja2